Do not wrap around when focusing another container
When focusing the next container to the right we may actually focus one
that is left of the current one if we already reached the right most
container.
This behavior is undesired. This patch adjusts the focusing logic to no
longer wrap around. If we reached the right most container and issue a
'focus right' command the focus will simply stay where it is. The
behavior when changing the focus in another direction (left/up/down) is
adjusted in a similar way.