Do not close unused workspaces
i3 takes dynamics a bit too far for my taste: with the deletion of every
unused workspace it quickly becomes non-obvious which workspaces numbers
are used. A quick glance at the workspace list no longer is enough to
determine which workspace you are on, you have to read the number in
addition.
This change removes some of those dynamics by no longer causing i3 to
remove unused workspaces. Once created, a workspace is there to stay.