Create lower number workspaces as needed
It is very annoying that when switching to a workspace that this
workspace is created if it does not exist but not the ones with a lower
number.
This change adds a new function, workspace_select, that can be used
instead of workspace_show_by_name, for example, to show a workspace,
creating it and its lower number peers if necessary.