Allow for creation of new workspace when selecting next one
Until now, i3 when asked to select the next workspace would cycle
through the existing workspaces. That behavior is not the desired one.
Instead we would like it to switch to the workspace with the next higher
number or create it if it does not exist. A cycling through workspaces
is not needed.