i3bar: Draw indication whether workspaces have windows or not
Awesome has the nice feature that it will indicate on each workspace
whether it contains windows or not by drawing a small rectangle along
with the workspace indication in the bar.
This change sets out to make i3 behave the same way. To accomplish this
task we introduce a new property, 'has_children', to the workspaces IPC
command that is indicative of whether a workspace has windows or not. We
then adjust i3 to inform IPC clients of workspace changes whenever a new
windows is created, destroyed, or moved to a different workspace.