Rework resizing logic
i3's logic for resizing, although understandably logical, is unintuitive
for users of Awesome, tmux, or vim. In particular, the way i3 resizes
cons is by finding the parent and then resizing all children such that
the currently active one's percentage increase or decreases (depending
on whether to grow or shrink it). The remaining con's percentage on the
overall width/height is then distributed equally.
The aforementioned other programs, however, take a different approach.
In essence they shift the border between the currently active con (or
whatever term they use) and the one next to it (i.e., right or below,
depending on the direction). If no next neighbour is found then the left
one is taken and the direction inversed.
This change implements the logic of the latter for i3.