Rework handling of multi monitor setups
Handling of multi monitor setups in i3 is quite different from that in
awesome and other window managers. In i3 the idea is that different
workspaces can reside on different monitors ("outputs" in i3 jargon).
Although that may work in scenarios where not many named workspaces are
in use, it is not a feasible approach when solely using numbered
workspaces in a dynamically created fashion.
To address this issue, this change reworks the logic slightly. Instead
of numbering workspaces continuously over monitors we introduce
boundaries: each monitor gets workspaces that start with the number one.
The remaining logic stays the same at a high level but a couple of
adjustments have to be made internally to honor outputs when looking up
workspaces.