Do not wrap around when switching outputs
When switching to another output i3 wraps around. For example, if we are
already focused on the right most output and issue a command to focus
the next output to the right it will actually start over with the left
most one.
This behavior is undesired. This patch adjusts the output changing logic
to no longer wrap around. If we reached the right most container and
issue a 'focus output right' command the focus will simply stay where it
is. The behavior when changing the output in another direction
(left/up/down) is adjusted in a similar way.