i3bar: Correctly focus output when handling a workspace mouse click
In a multi monitor setup, there are two problems when it comes to
selecting a workspace through the i3bar via a mouse click. First,
when the user selects a workspace on the i3bar on the output (i.e.,
monitor) that does not contain the currently focused workspace, this
workspace will not receive the focus if there is no application running
on it.
Second, in the same scenario a workspace switch can happen on the wrong
output, provided the workspace's name does not uniquely identify a
single object.
This change addresses both problems by reworking the command that i3bar
sends to switch workspaces. In particular, it prefixes the 'workspace
<ws-name>' part with an 'focus output <output-name>' command, which
takes care of focusing the correct output. By doing so, the subsequent
'workspace' command will also find the correct workspace in case of
duplicated workspace names because the search starts on the currently
focused output.
Note that both problems are not exposed on a vanilla i3. Only if i3 is
slightly modified to allow empty workspaces and duplicated workspace
names can this be an issue.