Correctly sort in workspaces when disabling outputs
When an output (i.e., monitor) that is managed by i3 is disabled, all
workspaces get moved to the "first" output. That becomes a problem with
the existing logic in the context of duplicated workspace numbers as
the number is used to determine where the insertion should happen.
This change addresses the problem by by adjusting the number (and name)
of the workspace before the insertion takes place in a way that existing
workspaces from the old output are appended to the list of existing
workspaces.