Honor current output in startup ID infrastructure
Currently, the startup ID infrastructure (used for 'exec' statements
without the --no-startup-id switch) does not honor the current output.
This can lead to problems in scenarios where workspace names are
ambiguous. For example, in a multi monitor setup an application started
on workspace 3 on the second monitor may actually spawn on workspace 3
on the first monitor, provided that the first monitor is currently
containing the focused container.
This change adjusts the startup module to store the current output name
along with the workspace name, which resolves this problem.