Close empty workspaces on i3 restart
The change to no longer destroy empty workspaces combined with i3's
ability to create them easily dynamically can lead to situations where
the user created undesired workspaces or no longer needs some of them.
Having no operation at all for cleaning up is not an ideal state to be
in.
To that end, this change adds functionality to destroy empty workspaces
again. As before we preserve the invariant that all workspaces (empty or
not) before the last non-empty one or the focused one (whichever is
further in the back) are left untouched.
In order to trigger the clean we piggy back on the i3 restart
functionality. Despite it being a relatively heavy operation (but still
reasonably fast), hooking in there is a very elegant way of achieving
this goal. Implementing a new command, as another possible approach, for
example, just leads to greater divergence from vanilla i3 in terms of
the configuration file.