Honor output in workspace_auto_back_and_forth logic
Similar to the startup ID logic, the workspace_auto_back_and_forth
functionality is not written taking the output of a workspace into
consideration. That can be a problem when workspace names are
ambiguous. In addition to that, the previous workspace is remembered on
a global scale, not on a per-output basis.
To fix both problems, this change causes i3 to store the last active
workspace on a per output basis. This way we can switch back and forth
between workspaces on the individual outputs, making the behavior more
intuitive for our use case.