Honor output in workspace_auto_back_and_forth logic
Similar to the startup ID logic, the workspace_auto_back_and_forth
functionality is not written taking the output of a workspace into
consideration. That can be a problem when workspace names are
ambiguous.
To fix this problem, this change stores not only the name of the
workspace that was focused last but also the name of the output it
belongs to. This way, the correct workspace can be looked up even if its
name is ambiguous.