Stop removing empty workspaces when an output is disabled
If an output is disabled, i3 gets rid of empty workspaces on it, while
migrating all others to a different output. This behavior is kind of
what one would want, at least in a multi-monitor setup. In a single
monitor setup, however, that causes workspaces to be rearranged every
time the last output is disabled (and eventually re-enabled), which is
not a great experience.
Unfortunately, it is not always possible to distinguish between single
and multi monitor setups. A monitor may be recognized even if a physical
output is merely used for digital audio.
To cater better to the single monitor case, with this change we adjust
the logic to stop removing empty workspaces altogether.