Stop removing empty workspaces when an output is disabled
If an output is disabled i3 gets rid of empty workspaces on it, while
"merging" all others into a different output. This behavior is kind of
what one would want, especially in a multi-monitor setup. In a single
monitor setup, however, that causes workspaces to be rearranged every
time the last output is disabled (and eventually re-enabled).
Unfortunately, it is not always possible to distinguish ...